second commit

+++ src/app/employee-dashboard/employee-dashboard.component.ts
@@ -1,5 +1,5 @@
 import { Component, OnInit } from '@angular/core';
-import { FormBuilder, FormGroup } from '@angular/forms';
+import { FormBuilder, FormGroup, Validators } from '@angular/forms';
 import { ApiService } from '../shared/api.service';
 import { EmployeeModel } from './employee-dashboard.model';
 
@@ -11,15 +11,15 @@ import { EmployeeModel } from './employee-dashboard.model';
 export class EmployeeDashboardComponent implements OnInit {
 
   formValue !: FormGroup;
+  isUpdated = false;
   employeeModelObj: EmployeeModel = new EmployeeModel();
   employeeData !: any;
   constructor(private formbuilder: FormBuilder, private api: ApiService) { }
 
   ngOnInit(): void {
     this.formValue = this.formbuilder.group({
-      id: [''],
-      userid: [''],
-      title: [''],
+      description: ['',[Validators.required]],
+      title: ['',[Validators.required,Validators.maxLength(20)]],
       status: ['']
     })
     this.getAllEmployee()
@@ -31,8 +31,12 @@ export class EmployeeDashboardComponent implements OnInit {
     this.employeeModelObj.userid = this.formValue.value.userid;
     this.employeeModelObj.title = this.formValue.value.title;
     this.employeeModelObj.status = this.formValue.value.status;
-
-    this.api.postEmploye(this.employeeModelObj)
+    
+    if(!this.formValue.valid){
+      alert("Enter Valid Title");
+      return ;
+    }
+    this.api.postEmploye(this.formValue.value)
     //console.log(this.employeeModelObj);
       .subscribe((res:any) => {
         console.log(res);
@@ -56,6 +60,10 @@ export class EmployeeDashboardComponent implements OnInit {
   }
   deleteEmployee(row : any){
     console.log(row)
+    let confirm = window.confirm("Are you sure you want to delete this employee data?")
+    if(!confirm){
+      return ;
+    }
     this.api.deleteEmploye(row.id)
     .subscribe(res=>{
       alert("Employee Deleted");
@@ -63,28 +71,41 @@ export class EmployeeDashboardComponent implements OnInit {
     })
   }
   onEdit(row: any){
+    this.isUpdated = true
     console.log(row)
     this.employeeModelObj.id = row.id;
-    this.formValue.controls['userid'].setValue(row.userid);
-    this.formValue.controls['id'].setValue(row.id);
+    this.formValue.controls['description'].setValue(row.description);
+    // this.formValue.controls['id'].setValue(row.id);
     this.formValue.controls['title'].setValue(row.title);
     this.formValue.controls['status'].setValue(row.status);
+    // this.formValue.reset();
     // this.profileForm.patchValue(row)
   }
   UpdateEmployeedetails(){
-    this.employeeModelObj.id = this.formValue.value.id;
-    this.employeeModelObj.userid = this.formValue.value.userid;
-    this.employeeModelObj.title = this.formValue.value.title;
-    this.employeeModelObj.status = this.formValue.value.status;
-
-    this.api.updateEmploye(this.employeeModelObj,this.employeeModelObj.id)
+    
+    // this.employeeModelObj.id = this.formValue.value.id;
+    // this.employeeModelObj.userid = this.formValue.value.userid;
+    // this.employeeModelObj.title = this.formValue.value.title;
+    // this.employeeModelObj.des = this.formValue.value.title;
+    // this.employeeModelObj.status = this.formValue.value.status;
+    if(!this.formValue.valid){
+      alert("something went wromg!");
+      return ;
+    }
+    this.api.updateEmploye(this.formValue.value,this.employeeModelObj.id)
     .subscribe(res=>{
       alert("Updated Successfully");
       let ref= document.getElementById('cancel')
         ref?.click();
+        
         this.formValue.reset();
+        console.log(this.formValue)
         this.getAllEmployee();
         
     })
+  }
+  formreset(){
+    this.isUpdated = false
+    this.formValue.reset();
   }
 }
